refactor(cart): align lib/cart with {index,server,client,action} convention
- Drop the re-export barrel lib/cart/index.ts and move the universal
  cartDiscountAmount into a real index.ts (server+client safe, no directive);
  delete amount.ts.
- Fold createCartServerHandlers() into lib/cart/server.ts and delete the
  four-line handlers.ts; route.ts and proxy.ts import cartHandlers from
  @/lib/cart/server.
- Import Hydrogen's CartProvider/useCart directly in context.tsx instead of a
  client wrapper around createCartComponents (which only reproduced the default
  exports).
- Merge the thin CartProviderWrapper/CartContextSync/useCartRender wrappers
  into components/cart/context.tsx; delete provider.tsx and context-sync.tsx.
