+++ apps/template/app/layout.tsx
@@ -8,9 +8,9 @@ import { Suspense } from "react";
 import { ActionBar } from "@/components/action-bar";
 import { AgentButton } from "@/components/agent/agent-button";
 import { AnalyticsComponents } from "@/components/analytics";
+import { CartProviderWrapper } from "@/components/cart/context";
 import { CartNotifications } from "@/components/cart/notifications";
 import { CartOverlayBridge } from "@/components/cart/overlay-bridge";
-import { CartProviderWrapper } from "@/components/cart/provider";
 import { Footer } from "@/components/footer";
 import { Nav } from "@/components/nav";
 import { SiteSchema } from "@/components/schema/site-schema";
