+++ apps/template/components/cart-page/summary.tsx
@@ -4,10 +4,10 @@ import { Loader2 } from "lucide-react";
 import { useEffect, useState } from "react";
 
 import { useCart } from "@/components/cart/context";
-import { useCartRender } from "@/components/cart/context-sync";
+import { useCartRender } from "@/components/cart/context";
 import { DiscountForm } from "@/components/cart/discount-form";
+import { cartDiscountAmount } from "@/lib/cart";
 import { prepareCheckoutAction } from "@/lib/cart/action";
-import { cartDiscountAmount } from "@/lib/cart/amount";
 import { cn, formatPrice } from "@/lib/utils";
 
 function CheckoutLink({
