--- apps/template/lib/cart/amount.ts
@@ -1,8 +1,0 @@
-import type { Cart } from "@/lib/types";
-
-export function cartDiscountAmount(cart: Cart): number {
-  return cart.discountAllocations.reduce(
-    (sum, a) => sum + parseFloat(a.discountedAmount.amount),
-    0,
-  );
-}
