--- apps/template/lib/cart/handlers.ts
@@ -1,4 +1,0 @@
-import "server-only";
-import { createCartServerHandlers } from "@shopify/hydrogen";
-
-export const cartHandlers = createCartServerHandlers();
